% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqtl_lib.R
\docType{package}
\name{eQTLpipeline_SystemGenetics}
\alias{eQTLpipeline_SystemGenetics}
\title{reduced pipeline of eQTL analysis for lecture}
\description{
A package to facilitate the full pipeline of cis eQTL analysis
}
\details{
\tabular{ll}{
Package: \tab eQTLpipeline\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2015-11-03\cr
License: \tab LGPL \cr
LazyLoad: \tab yes\cr
}
}
\examples{
#insertexample()
}
\references{

}
\keyword{package}
